refactor(server): Add preparations for avahi based mDNS plugin
Currently a fork of mdnsd library is used to register and discover servers on
the network. Since this fork is not maintained anymore a switch to another
library is necessary.
If avahi daemon is already running on the system, it can be used to register
the servers on network. Add mDNS build option and a copy of mDNS file as
preparation for an additional avahi based mDNS plugin.

Signed-off-by: Vasilij Strassheim <vasilij.strassheim@linutronix.de>